:refactor: 盤面表示関数を切り出し
